Fixes 3163
