+++ .publicApi/Microsoft.AI.Web.dll/Stable/PublicAPI.Unshipped.txt
@@ -1,0 +1,2 @@
+Microsoft.ApplicationInsights.Web.WebApplicationInsightsInitializer
+static Microsoft.ApplicationInsights.Web.WebApplicationInsightsInitializer.Initialize() -> void
