+++ examples/ClassicAspNetWebApp/applicationinsights.config
@@ -27,6 +27,6 @@
   <EnablePerformanceCounterCollectionModule>true</EnablePerformanceCounterCollectionModule>
   <AddAutoCollectedMetricExtractor>true</AddAutoCollectedMetricExtractor>
   <EnableDependencyTrackingTelemetryModule>true</EnableDependencyTrackingTelemetryModule>
-  <EnableRequestTrackingTelemetryModule>false</EnableRequestTrackingTelemetryModule>
+  <EnableRequestTrackingTelemetryModule>true</EnableRequestTrackingTelemetryModule>
   <ApplicationVersion>1.0.0</ApplicationVersion>
 </ApplicationInsights>
